Calculate individual species coexistence in regional scale

Calculate individual species coexistence in regional scale

Usage

IndivSpecies_regional_coex(
  df.TS.TE,
  time.slice,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE"
)

Arguments

df.TS.TE

a data frame object containing at least three columns. Species names, origination time and extinction time for each species.

time.slice

scalar indicating the time interval between consecutive time slices

round.digits

scalar indicating the number of digits for time of origination and time for extinction

species

character indicating the name of the column of the data frame containing the species name information

TS

character indicating the name of the columns of the data frame containing the information on origination time

TE

character indicating the name of the column of the data frame containing the information on extinction time

Value

A list with two elements:

df_IndivSpp_coexist

A data frame with species name, the number of coexistence for each sepecies in each time slice

mean_species_coexistence

A data frame with three column. The species name, the mean number of coexistence for each species and the variance considering all time slices