Compute regional clade Mean Pairwise Distance in a time series

Compute regional clade Mean Pairwise Distance in a time series

Usage

clade_regional_mpd(
  df.TS.TE,
  time.slice,
  trait,
  round.digits = 1,
  species = "species",
  TS = "TS",
  TE = "TE",
  compute.ses.mpd = TRUE,
  null.model = "taxa.labels",
  runs = 999,
  group = NULL,
  group.focal.compare = NULL,
  type.comparison = NULL
)

Arguments

runs